Fix a signed/unsigned comparison warning.

+++ tests/lib/fakewrite.c
@@ -66,7 +66,7 @@ fake_pwrite(int fd UNUSED, const void *data, size_t n, off_t offset)
         errno = ENOSPC;
         return -1;
     }
-    if (256 - offset < total)
+    if ((size_t) (256 - offset) < total)
         total = 256 - offset;
     memcpy(write_buffer + offset, data, total);
     return total;
